Rename actix-web package to actix-web-crate to simplify features
Unfortunately, cargo doesn't allow us to piggy-back on an existing
feature that is defined through an optional crate.

Therefore, to be able to use the name 'actix-web' for a feature,
we need to rename the actix-web dependency to something else.

Doing so allows us to drop the `with` prefixes for all features
which is more natural.